backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
playwright-stealth==1.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
celery==5.3.4
redis==5.0.1
pandas==2.1.3
openpyxl==3.1.2
alembic==1.12.1